pyproject.toml
# Copyright 2025 Canonical Ltd.
# See LICENSE file for licensing details.
[project]
name = "pyroscope-k8s"
version = "0.1"  # this is in fact irrelevant
requires-python = ">=3.12"


[project.optional-dependencies]
dev = [
    #INTEGRATION TESTS
    "jubilant",
    "pytest-jubilant>=0.6",
    "minio",
    "requests",
    "coordinated-workers>=2.0.0",
    "tenacity",
]


[tool.pytest.ini_options]
minversion = "6.0"
log_cli_level = "INFO"
markers = ["setup", "teardown"]


